requirements.txt
streamlit
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
pypdf
python-docx
python-dotenv
torch==2.11.0
torchvision==0.26.0
torchaudio==2.11.0